requirements.txt
dropbox==12.0.2
eel==0.17.0
pyAesCrypt==6.1.1
pyqrcode==1.2.1
python-dotenv==1.0.1
pyzbar==0.1.9
opencv_python==4.10.0
requests==2.32.3